rename aws-dynamo to aws-dynamodb
